chore: rename pkg/namespace-labeler to pkg/namespacelabeler
